feat: calculate franchise tag

+++ db/schema.sql
@@ -724,10 +724,10 @@ DROP TABLE IF EXISTS `seasons`;
 CREATE TABLE `seasons` (
   `lid` int(11) NOT NULL,
   `year` int(4) NOT NULL,
-  `fqb` int(3) DEFAULT NULL, -- qb franchise tag amount
-  `frb` int(3) DEFAULT NULL, -- rb franchise tag amount
-  `fwr` int(3) DEFAULT NULL, -- wr franchise tag amount
-  `fte` int(3) DEFAULT NULL, -- te franchise tag amount
+  `fqb` mediumint(3) unsigned DEFAULT NULL, -- qb franchise tag amount
+  `frb` mediumint(3) unsigned DEFAULT NULL, -- rb franchise tag amount
+  `fwr` mediumint(3) unsigned DEFAULT NULL, -- wr franchise tag amount
+  `fte` mediumint(3) unsigned DEFAULT NULL, -- te franchise tag amount
   UNIQUE KEY `season` (`lid`,`year`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
